# feat(web): improve session retrieval performance