--- a/apps/web/lib/core/i18n/i18n.utils.ts
+++ b/apps/web/lib/core/i18n/i18n.utils.ts
@@ -1,9 +1,10 @@
 import parser from "accept-language-parser";
 import type { IncomingMessage } from "http";
+import type { GetServerSidePropsContext } from "next";
+import { AUTH_OPTIONS } from "pages/api/auth/[...nextauth]";
 
+import { getServerSession } from "@calcom/lib/auth";
 import type { Maybe } from "@calcom/trpc/server";
-
-import { getSession } from "@lib/auth";
 
 import { i18n } from "../../../next-i18next.config";
 
@@ -15,9 +16,13 @@ export function getLocaleFromHeaders(req: IncomingMessage): string {
   return preferredLocale ?? i18n.defaultLocale;
 }
 
-export const getOrSetUserLocaleFromHeaders = async (req: IncomingMessage): Promise<string> => {
-  const prisma = (await import("@calcom/prisma")).default;
-  const session = await getSession({ req });
+export const getOrSetUserLocaleFromHeaders = async (
+  req: GetServerSidePropsContext["req"],
+  res: GetServerSidePropsContext["res"]
+): Promise<string> => {
+  const { default: prisma } = await import("@calcom/prisma");
+
+  const session = await getServerSession({ req, res, authOptions: AUTH_OPTIONS });
   const preferredLocale = getLocaleFromHeaders(req);
 
   if (session?.user?.id) {