--- a/apps/web/pages/api/auth/[...nextauth].tsx
+++ b/apps/web/pages/api/auth/[...nextauth].tsx
@@ -3,7 +3,7 @@ import { IdentityProvider } from "@prisma/client";
 import { readFileSync } from "fs";
 import Handlebars from "handlebars";
 import { SignJWT } from "jose";
-import type { Session } from "next-auth";
+import type { AuthOptions, Session } from "next-auth";
 import NextAuth from "next-auth";
 import { encode } from "next-auth/jwt";
 import type { Provider } from "next-auth/providers";
@@ -322,7 +322,7 @@ function isNumber(n: string) {
 
 const calcomAdapter = CalComAdapter(prisma);
 
-export default NextAuth({
+export const AUTH_OPTIONS: AuthOptions = {
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
   adapter: calcomAdapter,
@@ -656,4 +656,6 @@ export default NextAuth({
       return baseUrl;
     },
   },
-});
+};
+
+export default NextAuth(AUTH_OPTIONS);