--- a/apps/web/pages/api/book/event.ts
+++ b/apps/web/pages/api/book/event.ts
@@ -1,11 +1,12 @@
-import type { NextApiRequest } from "next";
+import type { NextApiRequest, NextApiResponse } from "next";
+import { AUTH_OPTIONS } from "pages/api/auth/[...nextauth]";
 
 import handleNewBooking from "@calcom/features/bookings/lib/handleNewBooking";
-import { getSession } from "@calcom/lib/auth";
+import { getServerSession } from "@calcom/lib/auth";
 import { defaultResponder } from "@calcom/lib/server";
 
-async function handler(req: NextApiRequest & { userId?: number }) {
-  const session = await getSession({ req });
+async function handler(req: NextApiRequest & { userId?: number }, res: NextApiResponse) {
+  const session = await getServerSession({ req, res, authOptions: AUTH_OPTIONS });
   /* To mimic API behavior and comply with types */
   req.userId = session?.user?.id || -1;
   const booking = await handleNewBooking(req, {