--- a/apps/web/pages/api/integrations/[...args].ts
+++ b/apps/web/pages/api/integrations/[...args].ts
@@ -1,9 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import type { Session } from "next-auth";
-import { AUTH_OPTIONS } from "pages/api/auth/[...nextauth]";
 
 import getInstalledAppPath from "@calcom/app-store/_utils/getInstalledAppPath";
-import { getServerSession } from "@calcom/lib/auth";
+import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
 import { deriveAppDictKeyFromType } from "@calcom/lib/deriveAppDictKeyFromType";
 import { revalidateCalendarCache } from "@calcom/lib/server/revalidateCalendarCache";
 import prisma from "@calcom/prisma";
@@ -43,7 +42,7 @@ const defaultIntegrationAddHandler = async ({
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   // Check that user is authenticated
-  req.session = await getServerSession({ req, res, authOptions: AUTH_OPTIONS });
+  req.session = await getServerSession({ req, res });
 
   const { args } = req.query;