--- a/apps/web/pages/api/trpc/[trpc].ts
+++ b/apps/web/pages/api/trpc/[trpc].ts
@@ -1,10 +1,12 @@
 /**
  * This file contains tRPC's HTTP response handler
  */
+import { AUTH_OPTIONS } from "pages/api/auth/[...nextauth]";
 import { z } from "zod";
 
+import { getServerSession } from "@calcom/lib/auth";
 import * as trpcNext from "@calcom/trpc/server/adapters/next";
-import { createContext } from "@calcom/trpc/server/createContext";
+import { createContext as createTrpcContext } from "@calcom/trpc/server/createContext";
 import { appRouter } from "@calcom/trpc/server/routers/_app";
 
 export default trpcNext.createNextApiHandler({
@@ -12,7 +14,11 @@ export default trpcNext.createNextApiHandler({
   /**
    * @link https://trpc.io/docs/context
    */
-  createContext,
+  createContext: ({ req, res }) => {
+    const sessionGetter = () => getServerSession({ req, res, authOptions: AUTH_OPTIONS });
+
+    return createTrpcContext({ req, res }, sessionGetter);
+  },
   /**
    * @link https://trpc.io/docs/error-handling
    */