… apps/web/lib/auth.ts
… apps/web/package.json
… apps/web/pages/apps/index.tsx
… apps/web/pages/auth/forgot-password/index.tsx
… apps/web/pages/auth/login.tsx
… apps/web/pages/auth/setup/index.tsx
… apps/web/pages/auth/sso/[provider].tsx
… apps/web/pages/event-types/[type]/index.tsx
… apps/web/pages/getting-started/[[...step]].tsx
… apps/web/pages/index.tsx
… apps/web/pages/settings/my-account/profile.tsx
… apps/web/pages/settings/security/password.tsx
… apps/web/pages/video/[uid].tsx
… apps/web/server/lib/ssr.ts
… apps/web/test/lib/getSchedule.test.ts
… packages/features/auth/SAMLLogin.tsx
… packages/features/auth/lib/ErrorCode.ts
… packages/features/auth/lib/ensureSession.ts
… packages/features/auth/lib/getServerSession.ts
… packages/features/auth/lib/getSession.ts
… packages/features/auth/lib/hashPassword.ts
… packages/features/auth/lib/identityProviderNameMap.ts
… packages/features/auth/lib/isPasswordValid.ts
… packages/features/auth/lib/next-auth-custom-adapter.ts
… packages/features/auth/lib/next-auth-options.ts
… packages/features/auth/lib/validPassword.ts
… packages/features/auth/lib/verifyPassword.ts
… packages/features/auth/package.json
… packages/features/ee/impersonation/lib/ImpersonationProvider.ts
… packages/features/ee/teams/api/upgrade.ts