@Mythie 2023-03-09
> Found a bug with a type that is inferred from the NextAuth options, will look into it shortly

Not a bug, just an editor mishap! All sorted