@Mythie 2023-03-10
> pretty big PR, 30 files touched. is it possible to split this up into smaller ones? If not, no biggie. we will review, may take a bit longer tho

Sadly not, but the changes for most files are the exact same where they replace `getSession({ req })` with the `getServerSession` alternative.

Hopefully a pretty chill review due to most of it being repeitive.