@Mythie 2023-03-10
Can I also suggest testing locally and verifying that the dev environment feels snappier than it previously was. Since pages/routes get recompiled it can be a bit harder to measure, if you want assistance with this I'd be happy to pair up.