@emrysal 2023-03-10
Hi @Mythie Thanks for this; looks all good to me - just wondering if you would have a look at our /ee folder, Particularly ImpersonationProvider and upgrade.ts, and see if these should also benefit from getServerSession ? 

I think you missed this as you weren't aware these could also be serverside only (I could be wrong)