@Mythie 2023-03-10
> Hi @Mythie Thanks for this; looks all good to me - just wondering if you would have a look at our /ee folder, Particularly ImpersonationProvider and upgrade.ts, and see if these should also benefit from getServerSession ?
> 
> I think you missed this as you weren't aware these could also be serverside only (I could be wrong)

These would benefit from it however it’s hard to get them said information as they are exported as full fledged handlers within their respective packages. 

Due to this I’ve opted to omit it as it would require a general restructuring.